De-duplicated RpcUtil
