+++ testkit/src/main/scala/org/bitcoins/util/AsyncUtil.scala
@@ -1,0 +1,5 @@
+package org.bitcoins.util
+
+trait AsyncUtil extends org.bitcoins.rpc.util.AsyncUtil
+
+object AsyncUtil extends AsyncUtil
